Fix various minor warnings generated by clang -Weverything

+++ src/agent.cc
@@ -148,16 +148,16 @@ void Agent::start(const char *transport_override){
 	GenericStruct *global=GenericManager::get()->getRoot()->get<GenericStruct>("global");
 		list<string> transports = global->get<ConfigStringList>("transports")->read();
 	//sofia needs a value in millseconds.
-	int tports_idle_timeout = 1000 * global->get<ConfigInt>("idle-timeout")->read();
+	unsigned int tports_idle_timeout = 1000 * (unsigned int)global->get<ConfigInt>("idle-timeout")->read();
 	bool mainPeerCert = global->get<ConfigBoolean>("require-peer-certificate")->read();
 	string mainTlsCertsDir = global->get<ConfigString>("tls-certificates-dir")->read();
-	int t1x64=global->get<ConfigInt>("transaction-timeout")->read();
+	unsigned int t1x64=(unsigned int)global->get<ConfigInt>("transaction-timeout")->read();
 	mainTlsCertsDir = absolutePath(currDir, mainTlsCertsDir);
 
 	SLOGD << "Main tls certs dir : " << mainTlsCertsDir;
 
 	nta_agent_set_params(mAgent,NTATAG_SIP_T1X64(t1x64),TAG_END());
-	
+
 	if (transport_override){
 		transports=ConfigStringList::parse(transport_override);
 	}
@@ -193,7 +193,7 @@ void Agent::start(const char *transport_override){
 			err=nta_agent_add_tport(mAgent,
 									(const url_string_t*) url,
 									TPTAG_CERTIFICATE(keys.c_str()),
-									TPTAG_TLS_VERIFY_PEER(peerCert),  
+									TPTAG_TLS_VERIFY_PEER(peerCert),
 									NTATAG_TLS_RPORT(1),
 									TPTAG_IDLE(tports_idle_timeout),
 									TAG_END());
@@ -209,7 +209,7 @@ void Agent::start(const char *transport_override){
 		}
 		su_home_deinit(&home);
 	}
-	
+
 	tport_t *primaries=tport_primaries(nta_agent_tports(mAgent));
 	if (primaries==NULL) LOGF("No sip transport defined.");
 	su_md5_t ctx;
@@ -257,30 +257,30 @@ void Agent::start(const char *transport_override){
 			//LOGD("\tDetected %s preferred route to %s", isIpv6 ? "ipv6":"ipv4", prefUrl);
 		}
 	}
-	
+
 	if (mPublicIpV4.empty() && mPreferredRouteV4) mPublicIpV4=mPreferredRouteV4->url_host;
 	if (mPublicIpV6.empty() && mPreferredRouteV6) mPublicIpV6=mPreferredRouteV6->url_host;
-	
+
 	if (mRtpBindIp.empty() && mPreferredRouteV4) {
 		mRtpBindIp=mPreferredRouteV4->url_host;
 	}
 	if (mRtpBindIp6.empty() && mPreferredRouteV6) {
 		mRtpBindIp6=mPreferredRouteV6->url_host;
 	}
-	
+
 	if (mRtpBindIp.empty()) mRtpBindIp="0.0.0.0";
 	if (mRtpBindIp6.empty()) mRtpBindIp6="::0";
-	
+
 	char digest[(SU_MD5_DIGEST_SIZE*2)+1];
 	su_md5_hexdigest(&ctx,digest);
 	su_md5_deinit(&ctx);
 	digest[16]='\0';//keep half of the digest, should be enough
 	// compute a network wide unique id
 	mUniqueId = digest;
-	
+
 	LOGD("Agent public hostname/ip: v4:%s v6:%s",mPublicIpV4.c_str(), mPublicIpV6.c_str());
 	LOGD("Agent's _default_ RTP bind ip address: v4:%s v6:%s",mRtpBindIp.c_str(),mRtpBindIp6.c_str());
-	
+
 	char prefUrl4[256]={0};
 	char prefUrl6[256]={0};
 	if (mPreferredRouteV4) url_e(prefUrl4,sizeof(prefUrl4),mPreferredRouteV4);
@@ -292,7 +292,7 @@ void Agent::start(const char *transport_override){
 Agent::Agent(su_root_t* root):mBaseConfigListener(NULL), mTerminating(false){
 	mHttpEngine = nth_engine_create(root, NTHTAG_ERROR_MSG(0), TAG_END());
 	GenericStruct *cr = GenericManager::get()->getRoot();
-	
+
 	EtcHostsResolver::get();
 
 	mModules.push_back(ModuleFactory::get()->createModuleInstance(this, "SanityChecker"));
@@ -504,7 +504,7 @@ bool Agent::Network::isInNetwork(const struct sockaddr *addr) const {
 string Agent::Network::print(const struct ifaddrs *ifaddr) {
 	stringstream ss;
 	int err;
-	int size = (ifaddr->ifa_addr->sa_family == AF_INET) ? sizeof(struct sockaddr_in) : sizeof(struct sockaddr_in6);
+	unsigned int size = (ifaddr->ifa_addr->sa_family == AF_INET) ? sizeof(struct sockaddr_in) : sizeof(struct sockaddr_in6);
 	char result[IPADDR_SIZE];
 	ss << "Name: " << ifaddr->ifa_name;
 
@@ -536,9 +536,9 @@ int Agent::countUsInVia(sip_via_t *via) const {
 
 bool Agent::isUs(const char *host, const char *port, bool check_aliases) const {
 	char *tmp = NULL;
-	int end;
+	size_t end;
 	tport_t *tport=tport_primaries(nta_agent_tports(mAgent));
-	
+
 	//skip possibly trailing '.' at the end of host
 	if (host[end = (strlen(host) - 1)] == '.') {
 		tmp = (char*) alloca(end+1);
@@ -617,7 +617,7 @@ template <typename SipEventT>
 inline void Agent::doSendEvent
 (shared_ptr<SipEventT> ev, const list<Module *>::iterator &begin, const list<Module *>::iterator &end) {
 	#define LOG_SCOPED_EV_THREAD(ssargs, key) LOG_SCOPED_THREAD(key, ssargs->getOrEmpty(key));
-	
+
 	auto ssargs=ev->getMsgSip()->getSipAttr();
 	LOG_SCOPED_EV_THREAD(ssargs, "from.uri.user");
 	LOG_SCOPED_EV_THREAD(ssargs, "from.uri.domain");
@@ -625,7 +625,7 @@ inline void Agent::doSendEvent
 	LOG_SCOPED_EV_THREAD(ssargs, "to.uri.domain");
 	LOG_SCOPED_EV_THREAD(ssargs, "method_or_status");
 	LOG_SCOPED_EV_THREAD(ssargs, "callid");
-	
+
 
 	for (auto it = begin; it != end; ++it) {
 		ev->mCurrModule = (*it);
@@ -635,7 +635,7 @@ inline void Agent::doSendEvent
 	}
 	if (!ev->isTerminated() && !ev->isSuspended()) {
 		LOGA("Event not handled");
-	}	
+	}
 }
